books and resources
Quantitative finance involves the use of mathematical models and computational techniques to analyze financial markets and securities, aiming to predict market movements and optimize investment strategies. It relies heavily on complex algorithms, statistical methods, and numerical analysis to make informed decisions on trading, asset management, risk management, and more.

Starting with a foundation in Python is essential for quants, as it's a versatile programming language widely used for data analysis, machine learning, and algorithmic trading due to its powerful libraries like NumPy, pandas, and scikit-learn. Machine learning and AI are integral for developing predictive models and automating trading strategies, enabling the analysis of vast datasets to identify patterns and make predictions. Data science skills are crucial for handling, processing, and analyzing the large volumes of data that drive quantitative analysis. Lastly, a background in physics can be surprisingly beneficial due to its focus on complex systems, modeling, and statistical analysis, providing a unique perspective on problem-solving and quantitative reasoning in finance.